% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doh-get.R
\name{doh_get}
\alias{doh_get}
\title{Make a DoH Request (GET/REST)}
\usage{
doh_get(name, type = "a", extra_params = list(),
  service_path = "https://9.9.9.9/dns-query")
}
\arguments{
\item{name}{name to query for}

\item{type}{DNS query type (defaults to "\code{A}")}

\item{extra_params}{any special \code{GET} query parameter needed for a given server API endpoint.
this should be a named \code{list}.}

\item{server_path}{full URL path to the DoH server quer endpoint (defaults to Quad9).}
}
\value{
\code{NULL} (if the query failed) or a \code{data.frame} (tibble)
}
\description{
Issue a \code{GET} REST API query of type \code{type} for \code{name} to the
DoH endpoint specified at \code{server_path}.
}
\examples{
doh_get("rud.is", "A")
}
\references{
\url{https://tools.ietf.org/id/draft-ietf-doh-dns-over-https-05.html}
}
