% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playdoh-package.R
\docType{package}
\name{playdoh}
\alias{playdoh}
\alias{playdoh-package}
\title{Make 'DNS over HTTPS' Queries}
\description{
\itemize{
\item RFC: \url{https://tools.ietf.org/id/draft-ietf-doh-dns-over-https-05.html}
\item URL: \url{https://gitlab.com/hrbrmstr/playdoh}
\item BugReports: \url{https://gitlab.com/hrbrmstr/playdoh/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
